- adds main function to most scripts
For all .py files that are entry points (can be invoked as separate scripts via CLI commands) the whole code is wrapped in a main function. This has the advantages that later on code can be separated out into their own functions and these functions can be placed below the previous code. So the overall reading flow of the code is from top to bottom like prose. Without this pattern functions would need to be defined above the invocation.
- also deletes unnecessary author comments at the top of the file